Compilation fixes for xlc
git-svn-id: file:///home/roberto/Git/svn@3659 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: db0fbdb166c99c2b4dbcb97b48fe5f1fb94d5c35

+++ trunk/xmipp/libraries/reconstruction/reconstruct_fourier.cpp
@@ -197,7 +197,7 @@ void Prog_RecFourier_prm::produce_Side_info()
     kernel.setXmippOrigin();
     FOR_ALL_ELEMENTS_IN_MATRIX3D(kernel)
     {
-        double r=sqrt(k*k+i*i+j*j);
+        double r=sqrt((double)(k*k+i*i+j*j));
         if (r>=blob.radius) kernel(k,i,j)=0;
         else kernel(k,i,j)=blob_table(ROUND(r*iDelta));
     }
@@ -970,7 +970,7 @@ void Prog_RecFourier_prm::finishComputations( FileName out_name )
     pad_relation = (pad_relation * pad_relation * pad_relation);
     FOR_ALL_ELEMENTS_IN_MATRIX3D(Vout())
     {
-        double factor = Fourier_blob_table(ROUND(sqrt(k*k+i*i+j*j)
+        double factor = Fourier_blob_table(ROUND(sqrt((double)(k*k+i*i+j*j))
                                                  *iDeltaFourier));
         Vout(k,i,j) /= (factor/pad_relation);
     }
